docker/image/app/requirements.txt
Redis
Flask
django
djangorestframework
mysqlclient
